MacOS/RAG_SQL_Gen/requirements.txt
unstructured # Document loading
chromadb # Vector storage
openai # For embeddings and LLM
tiktoken  # For embeddings
streamlit
langchain_openai
langchain_core
langchain_community